pyproject.toml
[project]
name = "veil"
version = "0.1.0"
description = "Personal modular platform - microkernel architecture"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "bcrypt>=5.0.0",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
    "starlette>=1.0.0",
    "alembic>=1.18.4",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "python-multipart>=0.0.26",
    # Aliyun OSS SDK
    "oss2>=2.18",
    # Playwright (crawler browser automation)
    "playwright>=1.40",
    # SSH/SFTP (cloud training remote deployment)
    "asyncssh>=2.14",
    # Async file operations
    "aiofiles>=23.0",
    # HTML parsing (crawler link extraction)
    "beautifulsoup4>=4.12",
    # SQLite async driver (dev default)
    "aiosqlite>=0.20",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.2",
    "ruff>=0.15.11",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
